\name{availableCountry}
\alias{availableCountry}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
SEARCH AVAILABLE CODES AND NAMES
}
\description{
Show available country names, ISO codes, WEO country codes, WEO subject codes for \code{searchWEO()}.
}
\usage{
availableISOCode()
availableCountry()
availableWEOCountryCode()
availableWEOSubjectCode()
}
\author{
Ming-Jer Lee <mingjerli@gmail.com>
}
\examples{
availableISOCode()
availableCountry()
availableWEOCountryCode()
availableWEOSubjectCode()
}