\name{rWEO-package}
\alias{rWEO-package}
\alias{rWEO}
\docType{package}
\title{
ACCESS LATEST IMF WORLD ECONOMIC OUTLOOK (WEO) DATA IN R
}
\description{
The rWEO package allows R to directly access World Economic Outlook data.
World Economic Outlook is basically a survey conducted and published by the International Monetary Fund. 
It is published twice and partly updated 3 times a year. 
It portrays the world economy in the near and medium context (basically 4 years). 
WEO forecasts include the macroeconomic indicators, such as GDP, inflation, current account and fiscal balance of more than 180 countries around the globe. It also deals with major economic policy issues.
}
\details{
\tabular{ll}{
Package: \tab rWEO\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2014-10-21\cr
License: \tab GPL-3\cr
}
}
\author{
Ming-Jer Lee

Maintainer: Ming-Jer Lee <mingjerli@gmail.com>
}
\references{
International Monetary Fund:
http://www.imf.org/
}
\keyword{ package }
\seealso{
}
\examples{
}
